move/merge test utility funcs in a '/test' and '/test/olm' pkgs
this avoids having multiple packages with test utility funcs,
especially in `pkg/test`
