+++ pkg/controller/cheinstallation/cheinstallation_controller_test.go
@@ -10,11 +10,9 @@ import (
 
 	"github.com/codeready-toolchain/toolchain-operator/pkg/apis"
 	"github.com/codeready-toolchain/toolchain-operator/pkg/apis/toolchain/v1alpha1"
-	"github.com/codeready-toolchain/toolchain-operator/pkg/test"
-	. "github.com/codeready-toolchain/toolchain-operator/pkg/test/k8s"
-	. "github.com/codeready-toolchain/toolchain-operator/pkg/test/olm"
-	. "github.com/codeready-toolchain/toolchain-operator/pkg/test/toolchain"
 	"github.com/codeready-toolchain/toolchain-operator/pkg/toolchain"
+	"github.com/codeready-toolchain/toolchain-operator/test"
+	. "github.com/codeready-toolchain/toolchain-operator/test/olm"
 
 	orgv1 "github.com/eclipse/che-operator/pkg/apis/org/v1"
 	"github.com/go-logr/logr"
