+++ test/olm/che_assertion.go
@@ -1,4 +1,4 @@
-package toolchain
+package olm
 
 import (
 	"context"
