+++ test/olm/conditions.go
@@ -1,4 +1,4 @@
-package toolchain
+package olm
 
 import (
 	"fmt"
